requirements.txt
Package                   Version
------------------------- -----------
aiofiles                  23.2.1
altair                    5.3.0
annotated-types           0.6.0
anyio                     4.3.0
attrs                     23.2.0
certifi                   2024.2.2
charset-normalizer        3.3.2
click                     8.1.7
contourpy                 1.2.1
cycler                    0.12.1
exceptiongroup            1.2.0
fastapi                   0.110.1
ffmpy                     0.3.2
filelock                  3.13.4
fonttools                 4.51.0
fsspec                    2024.3.1
gradio                    4.26.0
gradio_client             0.15.1
greenlet                  3.0.3
h11                       0.14.0
httpcore                  1.0.5
httpx                     0.27.0
huggingface-hub           0.22.2
idna                      3.7
importlib_resources       6.4.0
Jinja2                    3.1.3
jsonschema                4.21.1
jsonschema-specifications 2023.12.1
kiwisolver                1.4.5
markdown-it-py            3.0.0
MarkupSafe                2.1.5
matplotlib                3.8.4
mdurl                     0.1.2
numpy                     1.26.4
orjson                    3.10.0
packaging                 24.0
pandas                    2.2.2
pillow                    10.3.0
pip                       24.0
plotly                    5.21.0
psycopg2-binary           2.9.9
pydantic                  2.7.0
pydantic_core             2.18.1
pydub                     0.25.1
Pygments                  2.17.2
pyparsing                 3.1.2
python-dateutil           2.9.0.post0
python-multipart          0.0.9
pytz                      2024.1
PyYAML                    6.0.1
referencing               0.34.0
requests                  2.31.0
rich                      13.7.1
rpds-py                   0.18.0
ruff                      0.3.7
semantic-version          2.10.0
setuptools                69.1.0
shellingham               1.5.4
six                       1.16.0
sniffio                   1.3.1
SQLAlchemy                2.0.29
starlette                 0.37.2
tenacity                  8.2.3
tomlkit                   0.12.0
toolz                     0.12.1
tqdm                      4.66.2
typer                     0.12.3
typing_extensions         4.11.0
tzdata                    2024.1
urllib3                   2.2.1
uvicorn                   0.29.0
websockets                11.0.3
Werkzeug                  3.0.2
wheel                     0.42.0
zipp                      3.18.1
